+++ src/lib/usage/costCalculator.js
@@ -1,0 +1,56 @@
+/**
+ * Cost Calculator — extracted from usageDb.js (T-15)
+ *
+ * Pure function for calculating request cost based on model pricing.
+ * No DB interaction — pricing is fetched from localDb.
+ *
+ * @module lib/usage/costCalculator
+ */
+
+/**
+ * Calculate cost for a usage entry.
+ *
+ * @param {string} provider
+ * @param {string} model
+ * @param {Object} tokens
+ * @returns {Promise<number>} Cost in USD
+ */
+export async function calculateCost(provider, model, tokens) {
+  if (!tokens || !provider || !model) return 0;
+
+  try {
+    const { getPricingForModel } = await import("@/lib/localDb.js");
+    const pricing = await getPricingForModel(provider, model);
+    if (!pricing) return 0;
+
+    let cost = 0;
+
+    const inputTokens = tokens.input ?? tokens.prompt_tokens ?? tokens.input_tokens ?? 0;
+    const cachedTokens =
+      tokens.cacheRead ?? tokens.cached_tokens ?? tokens.cache_read_input_tokens ?? 0;
+    const nonCachedInput = Math.max(0, inputTokens - cachedTokens);
+    cost += nonCachedInput * (pricing.input / 1000000);
+
+    if (cachedTokens > 0) {
+      cost += cachedTokens * ((pricing.cached || pricing.input) / 1000000);
+    }
+
+    const outputTokens = tokens.output ?? tokens.completion_tokens ?? tokens.output_tokens ?? 0;
+    cost += outputTokens * (pricing.output / 1000000);
+
+    const reasoningTokens = tokens.reasoning ?? tokens.reasoning_tokens ?? 0;
+    if (reasoningTokens > 0) {
+      cost += reasoningTokens * ((pricing.reasoning || pricing.output) / 1000000);
+    }
+
+    const cacheCreationTokens = tokens.cacheCreation ?? tokens.cache_creation_input_tokens ?? 0;
+    if (cacheCreationTokens > 0) {
+      cost += cacheCreationTokens * ((pricing.cache_creation || pricing.input) / 1000000);
+    }
+
+    return cost;
+  } catch (error) {
+    console.error("Error calculating cost:", error);
+    return 0;
+  }
+}
